Move team.applemango.runnerbe.util.ActivityAlias to team.applemango.runnerbe.shared.android.util.ActivityAlias
