+++ features/register/snslogin/src/main/kotlin/team/applemango/runnerbe/feature/register/snslogin/SnsLoginActivity.kt
@@ -37,8 +37,8 @@ import team.applemango.runnerbe.shared.android.extension.setWindowInsetsUsage
 import team.applemango.runnerbe.shared.compose.extension.verticalInsetsPadding
 import team.applemango.runnerbe.shared.compose.theme.GradientAsset
 import team.applemango.runnerbe.shared.domain.extension.defaultCatch
-import team.applemango.runnerbe.util.DFMOnboardActivityAlias
-import team.applemango.runnerbe.util.MainActivityAlias
+import team.applemango.runnerbe.shared.android.util.DFMOnboardActivityAlias
+import team.applemango.runnerbe.shared.android.util.MainActivityAlias
 
 class SnsLoginActivity : ComponentActivity() {
